+++ services/web/server/tests/unit/with_dbs/03/test_users.py
@@ -37,16 +37,16 @@ from simcore_service_webserver.redis import (
     get_redis_user_notifications_client,
     setup_redis,
 )
-from simcore_service_webserver.redis_user_notifications import (
+from simcore_service_webserver.rest import setup_rest
+from simcore_service_webserver.security import setup_security
+from simcore_service_webserver.security_roles import UserRole
+from simcore_service_webserver.session import setup_session
+from simcore_service_webserver.user_notifications import (
     MAX_NOTIFICATIONS_FOR_USER,
     NotificationCategory,
     UserNotification,
     get_notification_key,
 )
-from simcore_service_webserver.rest import setup_rest
-from simcore_service_webserver.security import setup_security
-from simcore_service_webserver.security_roles import UserRole
-from simcore_service_webserver.session import setup_session
 from simcore_service_webserver.users import setup_users
 from simcore_service_webserver.users_handlers import _get_user_notifications
 from simcore_service_webserver.users_models import ProfileGet
